requirements.txt
djangorestframework-simplejwt[crypto]==5.2.2
